CV_tailor/requirements.txt
google-genai>=1.0.0
playwright>=1.40.0
python-telegram-bot>=21.0
